requirements.txt
pandas==1.0.5
SQLAlchemy==1.3.18
tqdm==4.48.2
python-dotenv==0.14.0
python_dateutil==2.8.1
scikit_learn==0.23.2
pipreqs==0.4.10
